New Scripts and code clean up.
Implemented TimerScript and MenuSelectionScript. Cleaned up code in the ScoringSystemScript due to repeat code that Copilot AI snuck in without my permission. Still getting some errors in unity with the RandomNumberGenerator but I will hold off on fixing it until I've started Script Implementation  to see if it will work then or not.