Updates and fixes to Script Implementation.
All scripts should be working properly with their UI elements. Added a proper method for the Runs to end, with a pop up opening displaying the final score and its multiplier, when closed it returns to the Main Menu scene. Finished implementing TimerScript (ties in to the End Run changes). If everything is working alright,Monday 26th will begin Program Testing.