Final Unbuilt version of Speed-Math.
Tons of debugging, but it works fine. Had to retweak the UI (Apparently I forgot to fix UI Scaling, so it looked horrible until now...well worse than it does now. I am NOT TOUCHING the UI again. It works perfectly fine).

Played it a bit myself and its suprsiingly fun if not difficult. Going to have some friends test it out for me and see if it needs a little tweaking....I HOPE THEY DONT BREAK SHIT.